move code to the config class
