+++ src/ddqa/models/jira.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: MIT
 from __future__ import annotations
 
+from collections.abc import Iterable
 from datetime import datetime
 
 from pydantic import BaseModel, Field, HttpUrl, validator
@@ -39,7 +40,7 @@ class JiraIssue(BaseModel):
 
 class JiraConfig(BaseModel):
     jira_server: HttpUrl
-    members: dict[str, str]
+    members: dict[str, str] = Field()
 
     @validator('members')
     def check_members(cls, v):  # noqa: N805
@@ -48,3 +49,22 @@ class JiraConfig(BaseModel):
             raise ValueError(message)
 
         return v
+
+    def get_jira_user_id_from_github_user_id(self, github_user_id: str) -> str | None:
+        return self.members.get(github_user_id)
+
+    def get_jira_user_ids_from_github_user_ids(self, github_user_ids: Iterable[str]) -> set[str]:
+        res = set()
+
+        for gh_id in github_user_ids:
+            if jira_id := self.get_jira_user_id_from_github_user_id(gh_id):
+                res.add(jira_id)
+
+        return res
+
+    def get_github_user_id_from_jira_user_id(self, jira_user_id: str) -> str | None:
+        for github_id, jira_id in self.members.items():
+            if jira_id == jira_user_id:
+                return github_id
+
+        return None
