+++ src/ddqa/utils/jira.py
@@ -196,25 +196,6 @@ class JiraClient:
 
         return new_issue
 
-    def get_github_user_id_from_jira_user_id(self, jira_user_id: str) -> str | None:
-        for github_id, jira_id in self.config.members.items():
-            if jira_id == jira_user_id:
-                return github_id
-
-        return None
-
-    def get_jira_user_id_from_github_user_id(self, github_user_id: str) -> str | None:
-        return self.config.members.get(github_user_id)
-
-    def get_jira_user_ids_from_github_user_ids(self, github_user_ids: Iterable[str]) -> set[str]:
-        res = set()
-
-        for gh_id in github_user_ids:
-            if jira_id := self.get_jira_user_id_from_github_user_id(gh_id):
-                res.add(jira_id)
-
-        return res
-
     async def __get_transitions(self, client: ResponsiveNetworkClient, issue: JiraIssue) -> None:
         issue_types = self.__transitions.setdefault(issue.project, {})
         if issue.type in issue_types:
